requirements.txt
Flask==2.3.3
Flask-Mail==0.9.1
Flask-CORS==4.0.0
gunicorn==21.2.0
pyzbar==0.1.9
python-dotenv==1.0.0
bcrypt==4.0.1
qrcode==7.4.2
pymongo==4.5.0
Pillow==10.0.0
joblib==1.3.2
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
opencv-python-headless==4.8.0.76
Werkzeug==2.3.7
